+++ portability-spi-transfer/src/main/java/org/dataportabilityproject/spi/transfer/TransferServiceProviderRegistry.java
@@ -1,0 +1,15 @@
+package org.dataportabilityproject.spi.transfer;
+
+/**
+ * Manages {@link TransferServiceProvider}s registered in the system.
+ */
+public interface TransferServiceProviderRegistry {
+
+    /**
+     * Returns the provider that supports the service id.
+     *
+     * @param serviceId the service id
+     */
+    TransferServiceProvider getServiceProvider(String serviceId);
+
+}
