+++ portability-transfer/build.gradle
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Defines the worker SPI. Extensions use this module to implement custom import/export functionality for the system.
- */
-package org.dataportabilityproject.spi.worker;
 
+dependencies {
+    compile project(':portability-spi-cloud')
+    compile project(':portability-spi-transfer')
+}
