+++ settings.gradle
@@ -2,9 +2,9 @@ rootProject.name = 'org.dataportabilityproject'
 
 include ':portability-client-types', ':portability-data-transfer-types'
 
-include ':portability-spi-gateway', ':portability-spi-cloud', ':portability-spi-worker'
+include ':portability-spi-gateway', ':portability-spi-cloud', ':portability-spi-transfer'
 
-include ':portability-gateway', ':portability-cloud', ':portability-worker-service'
+include ':portability-gateway', ':portability-cloud', ':portability-transfer'
 
 include ':extensions:cloud:portability-cloud-google', ':extensions:cloud:portability-cloud-microsoft'
